requirements.txt
langchain
langgraph
pydantic
python-dotenv
sqlite-utils
psycopg2-binary
